% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_continuous_plus.R
\name{cont_breaks_plus}
\alias{cont_breaks_plus}
\title{Find Pretty Breaks for Continuous Axes in ggplot While Ensuring End Labels}
\usage{
cont_breaks_plus(data, n = 5, buffer_frac = 0.05)
}
\arguments{
\item{data}{The single vector of strictly numeric data to find pretty breaks for. Required.}

\item{n}{A length-1 numeric value for the "target" number of breaks to create. Defaults to 5.}

\item{buffer_frac}{A length-1 numeric value corresponding to how close the end breaks must be to the end of the data for new breaks to not be added. Defaults to 0.05 (5\%).}
}
\value{
Returns a named lists containing breaks and limits to use for the \code{data} provided, to be used in a \verb{scale_*_continuous()} function in a \code{ggplot2} call.
}
\description{
This function attempts to find a set of breaks for a continuous variable such that there aren't too many breaks, the breaks are "pretty" values where possible, and breaks exist at or near to the range values of the variable. It is essential \code{pretty_breaks()} from the \code{scales} package except that it is more opinionated about needing breaks at or near both ends of the data range. If needed, the limits of the axis are expanded slightly to yield a new break just past the range of the data.
}
\examples{
cont_breaks_plus(iris$Sepal.Length)
}
