% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middleware.R
\docType{data}
\name{default_theme}
\alias{default_theme}
\title{An Opinionated Version of \code{ggplot2}'s Default \code{theme_gray} Plot Theme}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 136.
}
\usage{
default_theme
}
\value{
A object of class "theme_plus," which will subsequently trigger the \code{ggplot_add.theme_plus} S3 method.
}
\description{
A set of replacement arguments to \code{ggplot2}'s default \code{theme_gray()} plot theme, affecting elements such as axis lines and text, legend lines and rectangles, gridlines and plot backgrounds, and strip text and placement.
}
\keyword{datasets}
