% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onloads.R
\name{.onLoad}
\alias{.onLoad}
\title{geom_point_plus() shape registry}
\usage{
.onLoad(libname, pkgname)
}
\description{
Package-private environment used to store built-in and user-registered
point shapes for \code{geom_point_plus()}.
}
\keyword{internal}
