% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_mapped_constants}
\alias{extract_mapped_constants}
\title{Remove Constants From a List of Provided and Inherited Aesthetics}
\usage{
extract_mapped_constants(
  aes_obj,
  target_list,
  silence = FALSE,
  location = "mapping"
)
}
\description{
This is an internal convenience function that takes a list of mapped aesthetics and checks to see if any of them are actually constants (e.g., they were provided inside of aes() but are not mapping to a variable). If it finds anything, it removes them from mapping and instead ports them to a "target_list" and triggers a warning.
}
\keyword{internal}
