% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_plus.R
\docType{data}
\name{geom_plus_defaults}
\alias{geom_plus_defaults}
\title{Default settings for geometry layers created by \code{geom_plus()}}
\format{
A named list with elements like "point", "jitter", "boxplot", etc., corresponding to commonly used ggplot2 geometries. Use names(geom_plus_defaults) for a full list.
}
\usage{
geom_plus_defaults
}
\description{
A named list of default aesthetics used by \code{geom_plus()} to control styling of the resulting geometry layers.
}
\keyword{datasets}
