% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserFacingCoreFunctions.R
\name{geom_point_plus_shapes}
\alias{geom_point_plus_shapes}
\title{Demo plot showing geom_point_plus() shapes}
\format{
A ggplot object.
}
\usage{
geom_point_plus_shapes()
}
\description{
A prebuilt ggplot object displaying the nine custom point shapes.
}
