% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_plus.R
\name{get_theme_plus}
\alias{get_theme_plus}
\title{A Convenience Function to Access \code{ggplot.plus}'s Opinionated Default Theme}
\usage{
get_theme_plus(legend_pos = "top")
}
\arguments{
\item{legend_pos}{Character string (\code{"top"} or \code{"right"}).
Controls where the legend is placed and which theme settings are applied.
Defaults to \code{"top"}. Any other value will be interpretted as \code{"right"}.}
}
\value{
A \code{ggplot2::theme} object.
}
\description{
This function provides a convenient way to extract the package’s opinionated
default theme, including legend-position-specific settings.
Use the returned theme like you would any other \code{ggplot2} theme (see examples).
}
\examples{
library(ggplot2)

ggplot2::ggplot(mtcars, ggplot2::aes(wt, mpg, colour = factor(cyl))) +
  geom_plus(geom = "point") +
  get_theme_plus()                  # legend on top (default)

ggplot2::ggplot(mtcars, ggplot2::aes(wt, mpg, colour = factor(cyl))) +
  geom_plus(geom = "point") +
  get_theme_plus("right")           # legend on the right

ggplot2::theme_set(get_theme_plus()) # set theme to session default
}
\seealso{
\code{\link[=theme_plus]{theme_plus()}} for the full theme-building workflow.
}
