% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.plus-package.R
\docType{package}
\name{ggplot.plus-package}
\alias{ggplot.plus}
\alias{ggplot.plus-package}
\title{ggplot.plus: Elevating Ggplot2's Defaults Aesthetics and Behaviors}
\description{
A collection of ggplot2 enhancements for improved design defaults, accessibility, and usability.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MAISRC/ggplot_plus}
}

}
\author{
\strong{Maintainer}: Alex Bajcz \email{bajcz003@umn.com}

}
\keyword{internal}
