% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_plus.R
\name{ggplot_add.geom_plus}
\alias{ggplot_add.geom_plus}
\title{Add A geom_plus-generated Geometry to a ggplot}
\usage{
\method{ggplot_add}{geom_plus}(object, plot, object_name)
}
\arguments{
\item{object}{An object of class \code{geom_plus}, created by \code{geom_plus()}, containing user-provided arguments (if any).}

\item{plot}{A ggplot object to which the new geometry layer should be added.}

\item{object_name}{Internal name used by ggplot2 when adding the layer.}
}
\value{
A ggplot object with the new geometry layer added.
}
\description{
This method defines how objects of class \code{geom_plus()}, added by the geom_plus function, are added to a ggplot2 plot using the \code{+} operator.
It processes default aesthetics, handles user overrides, and ensures compatibility with ggplot2 layering.
}
