% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridlines_plus.R
\name{ggplot_add.gridlines_plus}
\alias{ggplot_add.gridlines_plus}
\title{Add A gridlines_plus-generated Geometry to a ggplot}
\usage{
\method{ggplot_add}{gridlines_plus}(object, plot, name)
}
\arguments{
\item{object}{An object of class \code{gridlines_plus}, created by \code{gridlines_plus()}, containing user-provided arguments (if any) or else pre-defined default values.}

\item{plot}{A ggplot object to which the new gridlines should be added.}

\item{name}{Internal name used by ggplot2 when adding the layer.}
}
\value{
A ggplot object with the new gridlines added.
}
\description{
This method defines how objects of class \code{gridlines_plus()}, added by the gridlines_plus function, are added to a ggplot2 plot using the \code{+} operator.
It considers both default values as well as user overrides for important gridlines features and ensures compatibility with ggplot2 layering.
}
