% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_continuous_plus.R
\name{ggplot_add.scale_x_cont_plus}
\alias{ggplot_add.scale_x_cont_plus}
\title{Add A scale_x_cont_plus-generated X axis Gradation to a ggplot}
\usage{
\method{ggplot_add}{scale_x_cont_plus}(object, plot, name)
}
\arguments{
\item{object}{An object of class \code{scale_x_continuous_plus}, created by \code{scale_x_continuous_plus()}, containing user-provided arguments (if any) or else pre-defined default values that find a set of "pretty" breaks that encompass the full range of values on the axis and that expand the limits of the axis slightly, if needed, to accomplish this.}

\item{plot}{A ggplot object to which the new x axis scale should be added.}

\item{name}{Internal name used by ggplot2 when adding the layer.}
}
\value{
A ggplot object with the x axis breaks and limits redefined.
}
\description{
This method defines how objects of class \code{scale_x_continuous_plus}, added by the function of the same name, are added to a ggplot2 plot using the \code{+} operator.
It ensures that the new, "pretty" breaks, now successfully anchored at or near the range values for the data in question, are added to the plot's x axis.
}
