% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_plus.R
\name{ggplot_add.theme_plus}
\alias{ggplot_add.theme_plus}
\title{Add A theme_plus-generated theme to a ggplot}
\usage{
\method{ggplot_add}{theme_plus}(object, plot, object_name)
}
\arguments{
\item{object}{An object of class \code{theme_plus}, created by \code{theme_plus()}, containing user-provided arguments (if any) and otherwise default values for many theme attributes.}

\item{plot}{A ggplot object to which the new theme will be applied}

\item{object_name}{Internal name used by ggplot2 when adding the theme.}
}
\value{
A ggplot object with the new theme applied.
}
\description{
This method defines how objects of class \code{theme_plus()}, added by the theme_plus function, are added to a ggplot2 plot using the \code{+} operator.
It applies user-specified overrides to sensible default values and ensures compatibility with ggplot2 layering.
}
