% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_plus.R
\name{ggplot_build.geom_plus_warnings}
\alias{ggplot_build.geom_plus_warnings}
\title{Build a ggplot With the Class "geom_plus_warnings".}
\usage{
\method{ggplot_build}{geom_plus_warnings}(plot)
}
\arguments{
\item{plot}{A ggplot object for which the checks should be performed.}
}
\value{
A built ggplot.
}
\description{
This method defines how objects of class \code{geom_plus_warnings}, created by the \code{ggplot_add.geom_plus()} function, are built into a ggplot2 plot.
The method is where various checks are performed to see if the user may be doing something "sub-optimal" design-wise that could be used to trigger an informative warning to steer better behaviors.
}
