% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaxis_title_plus.R
\name{ggplot_build.switcher}
\alias{ggplot_build.switcher}
\title{Build a ggplot With the Class "switcher".}
\usage{
\method{ggplot_build}{switcher}(plot)
}
\arguments{
\item{plot}{A ggplot object for which the y axis title should be moved.}
}
\value{
A ggplot with the class of "switched" to trigger the ggplot_gtable method of the same name and also with the \code{y_axis_switch_location} attribute set by the call to \code{y_axis_title_plus()}.
}
\description{
This method defines how objects of class \code{switcher}, created by the \code{ggplot_add.axis_switcher()} function, are built into a ggplot2 plot.
The method continues the process of rebuilding the ggplot with the y axis title moved to its new location within the gtable.
}
