% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaxis_title_plus.R
\name{ggplot_gtable.switched}
\alias{ggplot_gtable.switched}
\title{Finish a ggplot With the Class "switched".}
\usage{
\method{ggplot_gtable}{switched}(data)
}
\arguments{
\item{data}{A ggplot object with the class of "switched" for which the y axis title should be moved.}
}
\value{
A ggplot object compatible with \code{ggplot2}'s + command structure.
}
\description{
This method defines how objects of class \code{switched}, created by the \code{ggplot_build.switcher()} function, are finalized into a ggplot2 plot.
The method finishes the process of rebuilding the ggplot with the y axis title moved to its new location within the gtable.
}
