% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.plus-package.R
\docType{package}
\name{ggplot_plus-package}
\alias{ggplot_plus}
\alias{ggplot_plus-package}
\title{ggplot_plus: `ggplot_plus: An Opinionated, Universal-Design-Oriented Veneer For Ggplot2's Defaults}
\description{
A collection of enhancements to ggplot2's defaults, with a focus on universal accessibility and ease of use.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MAISRC/ggplot_plus}
}

}
\author{
\strong{Maintainer}: Alex Bajcz \email{bajcz003@umn.com}

}
\keyword{internal}
