% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes_plus.R
\name{palettes_reset}
\alias{palettes_reset}
\title{Restore Ggplot2 Palette Defaults (Undo \code{palettes_plus()})}
\usage{
palettes_reset()
}
\value{
(Invisibly) \code{NULL}.
}
\description{
Restores the ggplot2 palette-related \code{options()} keys to whatever they were
\strong{before} \code{palettes_plus()} was called. Called automatically
on package detach/unload but can also be called manually.
}
\examples{
palettes_plus()
# ... later ...
palettes_reset()

}
\seealso{
\code{\link[=palettes_plus]{palettes_plus()}}
}
