% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_continuous_plus.R
\name{scale_colour_continuous_plus}
\alias{scale_colour_continuous_plus}
\title{Find Pretty Breaks for a Continuous Color axis in ggplot While Ensuring End Labels}
\usage{
scale_colour_continuous_plus(
  ...,
  n = 5,
  buffer_frac = 0.05,
  thin_labels = FALSE
)
}
\arguments{
\item{...}{Standard inputs normally given to \code{scale_color_continuous()}. Must not include \code{breaks} or \code{limits} or an error will be returned, as the function attempts to circumvent the need to specify prettier breaks or appropriate limits.}

\item{n}{A length-1 numeric value for the "target" number of breaks to create. Defaults to 5. Passed to \code{cont_breaks_plus()} internally.}

\item{buffer_frac}{A length-1 numeric value corresponding to how close the end breaks must be to the end of the data for new breaks to not be added. Defaults to 0.05 (5\%). A length-1 numeric value for the "target" number of breaks to create. Defaults to 5. Passed to \code{cont_breaks_plus()} internally.}

\item{thin_labels}{Should every other label (starting with the second) be replaced with an empty string? Defaults to FALSE. Change to TRUE to enable. Useful for when the number of breaks/labels is high enough that the axis feels "over-labeled" in a way that might contribute to excess cognitive load.}
}
\value{
Returns a list of class \code{"scale_color_cont_plus"}, which will trigger the \code{ggplot_add} method by the same name to trigger the axis breaks reconfiguration.
}
\description{
This function is a alias for \code{scale_color_continuous_plus()}--see there for more information.
}
\examples{
ggplot2::ggplot(iris, ggplot2::aes(x=Sepal.Length, y=Petal.Length)) +
geom_plus(geom = "point", ggplot2::aes(color = Petal.Width)) +
scale_colour_continuous_plus()
}
