% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_point_plus.R
\docType{data}
\name{shapes.list}
\alias{shapes.list}
\title{A new palette of shapes to use in \code{ggplot2} scatterplots.}
\format{
An object of class \code{list} of length 9.
}
\usage{
shapes.list
}
\value{
A named list.
}
\description{
This named list contains elements with x and y coordinates that, when corrected, will draw new shapes to be used as points in \code{ggplot2} scatterplots. The piece attribute of each element dictates whether a region of the resulting shape will be solid or a "hole" in the final shape.
}
\keyword{datasets}
