% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_plus.R
\name{theme_plus}
\alias{theme_plus}
\title{Add a New Base Theme to ggplots With Elevated Defaults}
\usage{
theme_plus(...)
}
\arguments{
\item{...}{Other arguments to be passed along to the \code{theme()} (optional).}
}
\value{
List with the class "theme_plus", which will trigger the theme_plus method in ggplot_add.
}
\description{
Wrapper function for ggplot2's \code{theme()} function that still allows users to specify custom values for theme attributes but has default values for many attributes that are more likely to result in a graph that meets best practices for design aesthetics, usability, and accessibility.
}
\examples{
ggplot(iris, aes(x=Sepal.Length, y=Petal.Length)) + geom_plus(geom = "point") + theme_plus()
}
