% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaxis_title_plus.R
\name{translate_element}
\alias{translate_element}
\title{Convert Between a Couple of Different Ways of Referencing the Same Aesthetic}
\usage{
translate_element(el)
}
\arguments{
\item{el}{A list or list-like object containing the names of elements to be translated.}
}
\value{
A list of translated elements.
}
\description{
This is an internal convenience function that allows translation of the names of aesthetics like "colour" into ones like "col" used by \code{grid}'s functions.
}
