% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaxis_title_plus.R
\name{element_to_gpar}
\alias{element_to_gpar}
\title{Translate Between Ggplot's and Gpar's Attribute Names.}
\usage{
element_to_gpar(el)
}
\arguments{
\item{el}{A list or list-like object containing the names of elements to be translated.}
}
\value{
A list of translated elements.
}
\description{
This is an internal convenience function that matches up \code{ggplot2}'s aesthetics names with those expected by the \code{grid::gpar} function so that user-specified aesthetics get properly carried over into the final product.
}
