% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_plus.R
\name{geom_plus}
\alias{geom_plus}
\title{Generates Base Geoms With Elevated Defaults}
\usage{
geom_plus(geom, ...)
}
\arguments{
\item{geom}{The name of the geom being drawn. Corresponds to the portion of the geom_ function after the _, e.g., "point" for geom_point, "line" for geom_line, etc. Must be a length one character string, and must match an implemented geom. See names(geom_plus_defaults) for a list of these. Required input.}

\item{...}{Other arguments to be passed along to the geom_ function being called.}
}
\value{
List with the class "geom_plus", which will trigger the geom_plus method in ggplot_add.
}
\description{
Maps inputs to a base ggplot geom (e.g., geom_point or geom_line) but provides default values more likely to adhere to best practices around usability, design aesthetics, and accessibility.
}
\examples{
ggplot(iris, aes(x=Sepal.Length, y=Petal.Length)) + geom_plus(geom = "point")
}
