% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_point_plus.R
\name{geom_point2}
\alias{geom_point2}
\title{An backend version of \code{geom_point()} that can access distinctive shapes.}
\usage{
geom_point2(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  shapes = shapes.list,
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes(), as in \code{ggplot2::geom_point()}.}

\item{data}{The data to be displayed in this layer, as in \code{ggplot2::geom_point()}.}

\item{stat}{The statistical transformation to use on the data for this layer, as in \code{ggplot2::geom_point()}.}

\item{position}{A position adjustment to use on the data for this layer, as in \code{ggplot2::geom_point()}.}

\item{shapes}{A named list of custom shapes to be drawn in place of \code{ggplot2}'s standard palette of shapes.}

\item{...}{Other arguments passed on to layer()'s params argument, as in \code{ggplot2::geom_point()}.}

\item{na.rm}{Logical value controlling whether missing values should be removed from the data with a warning or silently, as in \code{ggplot2::geom_point()}.}

\item{show.legend}{Logical value controlling whether this layer should be included in the legends, as in \code{ggplot2::geom_point()}.}

\item{inherit.aes}{Logical for whether the default aesthetics should be overridden rather than combined with the provided aesthetics, as in \code{ggplot2::geom_point()}.}
}
\value{
A ggplot2 layer object.
}
\description{
This function behaves much like \code{ggplot2}'s \code{geom_point()} function except that it allows access to a palette of nine new shapes that vary in their openness, spikiness, and intersectionality, making them more easily distinguished. This function is not meant to be called directly--instead, \code{geom_point_plus()} calls and modifies this function and is the intended function for users.
}
