% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middleware.R
\name{geom_point_plus_shapes}
\alias{geom_point_plus_shapes}
\title{A new palette of distinct shapes made especially for \code{geom_point_plus}}
\usage{
geom_point_plus_shapes()
}
\value{
A named list.
}
\description{
Call this object to generate a \code{ggplot} showing the names and features of each of the shapes available to \code{ggplotplus::geom_point_plus}.
}
