% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_point_plus.R
\docType{data}
\name{geom_point_plus_shapes}
\alias{geom_point_plus_shapes}
\title{A new palette of shapes available to \code{geom_point_plus}}
\format{
An object of class \code{gg} (inherits from \code{ggplot}) of length 11.
}
\usage{
geom_point_plus_shapes
}
\value{
A named list.
}
\description{
Call this object to generate a \code{ggplot} showing the names and features of each of the shapes available to \code{ggplot.plus::geom_point_plus}.
}
\keyword{datasets}
