% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middleware.R
\docType{data}
\name{ggplotplus_shapes_list}
\alias{ggplotplus_shapes_list}
\title{A new palette of shapes to use in \code{ggplot2} scatterplots.}
\format{
An object of class \code{list} of length 9.
}
\usage{
ggplotplus_shapes_list
}
\value{
A named list.
}
\description{
This named list contains elements with x and y coordinates that, when corrected, will draw new shapes to be used as points in \code{ggplot2} scatterplots. The \code{piece} attribute of each element dictates whether a region of the resulting shape will be solid or a "hole" in the final shape.
}
\keyword{datasets}
