% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridlines_plus.R
\name{gridlines_plus}
\alias{gridlines_plus}
\title{Generates Subtle and Choice Gridlines on a ggplot}
\usage{
gridlines_plus(color = "gray90", linewidth = 1.2, linetype = "solid")
}
\arguments{
\item{color}{The color used for the gridlines. Defaults to "gray90". Must be a single character vector of length 1 corresponding to the name of a color.}

\item{linewidth}{Line width for the gridlines. Defaults to 1.2. Must be a single numeric value.}

\item{linetype}{Line type for the gridlines. Defaults to "solid." Must be a single character string value corresponding to an accepted linetype, such as "dotted" or "dashed".}
}
\value{
List with the class "gridlines_plus", which will trigger the gridlines_plus method in ggplot::ggplot_add.
}
\description{
This function adds, by default, subtle, light gray major gridlines to a ggplot graph only in directions mapped to continuous (and not discrete) variables.
}
\examples{
ggplot2::ggplot(iris, ggplot2::aes(x=Sepal.Length, y=Petal.Length)) +
geom_plus(geom = "point") +
gridlines_plus()
}
