% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes_plus.R
\name{palettes_plus}
\alias{palettes_plus}
\title{Makes the Viridis Family of Color Palettes the Default For ggplots}
\usage{
palettes_plus(palette = "D", begin = 0.28, end = 0.72)
}
\arguments{
\item{palette}{The viridis-family color palette to use. Defaults to "D", which corresponds to the viridis color palette. Must be a length-1 character string from "A" to "H" corresponding to the same codes used by \code{scale_color_viridis_d()} and similar functions to refer to the eight color families available in the viridis package.}

\item{begin}{The color value of the "low" end of the color palette. Defaults to 0.28, which is a medium-dark blue in the viridis palette. Must be a single numeric value between 0 and 1.}

\item{end}{The color value of the "high" end of the color palette. Defaults to 0.72, which is a medium-light green in the viridis palette. Must be a single numeric value between 0 and 1.}
}
\value{
Does not return an object.
}
\description{
This function coerces the default color palettes for the fill and color aesthetics of ggplots to be the \code{viridis} family of color palettes for the rest of the session by overriding defaults set inside of options(). This ensures the color palettes used with be accessible for those with diverse technological and occular needs.
}
\examples{
palettes_plus(); ggplot(iris, aes(x=Sepal.Length, y=Petal.Length, fill=Species)) + geom_plus(geom = "point")
}
