% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{translate_element}
\alias{translate_element}
\title{Convert Between a Couple of Different Ways of Referencing the Same Aesthetic}
\usage{
translate_element(el)
}
\description{
This is an internal convenience function that allows translation of the names of aesthetics like "colour" into ones like "col" used by \code{grid}'s functions.
}
\keyword{internal}
